Adds defcustom numpydoc-newline-after-opening-quotes
Adds `numpydoc-newline-after-opening-quotes` as a custom boolean variable with a default of `nil` to the `numpydoc` group.

This is then used in `numpydoc--insert-short-and-long-desc` to conditionally insert a new line as per [validation rule
GL01](https://numpydoc.readthedocs.io/en/latest/validation.html#built-in-validation-checks).

Currently though the short description (either `tmps` or the value from `read-string`) is not indented. Need to work out
why. Can't use...

```
                              (if numpydoc-newline-after-opening-quotes
                                  (insert "\n")
                                  (insert ""))
```

...since this inserts is _before_ the previous byte character (as per [documentation](https://www.gnu.org/software/emacs/manual/html_node/elisp/Insertion.html#index-insert))